+++ sdk/spring/spring-cloud-azure-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/implementation/jms/ServiceBusJmsConnectionFactoryFactory.java
@@ -4,7 +4,7 @@
 package com.azure.spring.cloud.autoconfigure.implementation.jms;
 
 import com.azure.spring.cloud.autoconfigure.implementation.jms.properties.AzureServiceBusJmsProperties;
-import com.azure.spring.cloud.autoconfigure.jms.ServiceBusJmsConnectionFactory;
+import com.azure.spring.jms.ServiceBusJmsConnectionFactory;
 import com.azure.spring.cloud.autoconfigure.jms.ServiceBusJmsConnectionFactoryCustomizer;
 import com.azure.spring.cloud.core.implementation.connectionstring.ServiceBusConnectionString;
 import org.apache.qpid.jms.policy.JmsDefaultPrefetchPolicy;