+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/implementation/aadb2c/security/AadB2cUrlTests.java
@@ -3,7 +3,6 @@
 package com.azure.spring.cloud.autoconfigure.implementation.aadb2c.security;
 
 
-import com.azure.spring.cloud.autoconfigure.implementation.aadb2c.security.AadB2cUrl;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;